+++ tests/client/track.py
@@ -14,7 +14,7 @@ class TestSpotifyTrack:
     def test_track_with_market(self, app_client):
         track = app_client.track(track_id, market='US')
         assert track.id == track_id
-        assert len(track.available_markets) == 0
+        assert track.available_markets is None
         assert track.is_playable is not None
 
     def test_track_no_market(self, app_client):
@@ -39,7 +39,7 @@ class TestSpotifyTrack:
     def test_tracks_with_market(self, app_client):
         tracks = app_client.tracks(track_ids, market='US')
         assert len(tracks) == len(track_ids)
-        assert all(len(track.available_markets) == 0 for track in tracks)
+        assert all(track.available_markets is None for track in tracks)
         assert all(track.is_playable is not None for track in tracks)
 
     def test_tracks_no_market(self, app_client):