+++ docs/src/release_notes.rst
@@ -9,6 +9,7 @@ Fixed
 *****
 - :ref:`client` - Document the need for at least one seed in
   :meth:`recommendations <Spotify.recommendations>` (:issue:`229`)
+- :ref:`client` - match new behavior of track markets (:issue:`231`)
 
 3.4.0 (2020-11-24)
 ------------------